[refactor] container object로 수정
